// src/app/shared/modules/popular-tags/types/popular-tags.interface.ts
export enum PopularTag {
  President = 'President',
  USA = 'USA',
  Freedom = 'Freedom',
  StandUp = 'StandUp',
  Comedy = 'Comedy',
  Life = 'Life',
  Simpsons = 'Simpsons',
  Beer = 'Beer',
  StoryOfMyLife = 'StoryOfMyLife',
  Shoe = 'Shoe',
  Prince = 'Prince',
  Ball = 'Ball',
  Pop = 'Pop',
  Music = 'Music',
  King = 'King',
  MeltDown = 'MeltDown',
  Dance = 'Dance',
  Rock = 'Rock',
  Nineties = 'Nineties',
  Rap = 'Rap',
  ThugLife = 'ThugLife',
  Winter = 'Winter',
  Xmas = 'Xmas',
  Holidays = 'Holidays',
  Soul = 'Soul',
}